Additional ACL response attributes
Include these attributes in the JSON response for access control lists.

* `etag`
* `kind`

Some [client libraries][1] expect them to be present, and fail to
deserialize the response otherwise.

We don't care too much about their values at this stage, so we just
picked a constant for `etag`. The value for `kind` matches what the real
GCS API returns.

[1]: https://crates.io/crates/google-cloud-storage